fixed void On<T1, T2>(
    public void On<T>(string methodName, Func<T, Task> handler) => inner.On(methodName, handler);
    public void On<T1, T2>(string methodName, Action<T1, T2> handler)
    {
        inner.On(methodName, [typeof(T1), typeof(T2)], (args) =>
        {
            handler((T1)args[0]!, (T2)args[1]!);
            return Task.CompletedTask;
        });
    }
