+++ OpenVPNGateMonitor/Services/DataGateOpenVpnManager/OpenVpnProxy/Hubs/DefaultHubConnectionFactory.cs
@@ -1,0 +1,21 @@
+using Microsoft.AspNetCore.SignalR.Client;
+using OpenVPNGateMonitor.Services.DataGateOpenVpnManager.OpenVpnProxy.Hubs.Interfaces;
+
+namespace OpenVPNGateMonitor.Services.DataGateOpenVpnManager.OpenVpnProxy.Hubs;
+
+internal sealed class DefaultHubConnectionFactory : IHubConnectionFactory
+{
+    public IHubConnectionProxy Create(string fullUrl, Func<Task<string?>> accessTokenProvider)
+    {
+        var connection = new HubConnectionBuilder()
+            .WithUrl(fullUrl, options => { options.AccessTokenProvider = accessTokenProvider; })
+            .WithAutomaticReconnect()
+            .ConfigureLogging(logging =>
+            {
+                logging.SetMinimumLevel(LogLevel.Debug);
+                logging.AddConsole();
+            })
+            .Build();
+        return new HubConnectionProxy(connection);
+    }
+}
