+++ OpenVPNGateMonitor/Services/DataGateOpenVpnManager/OpenVpnProxy/Hubs/Interfaces/IHubConnectionFactory.cs
@@ -1,0 +1,6 @@
+namespace OpenVPNGateMonitor.Services.DataGateOpenVpnManager.OpenVpnProxy.Hubs.Interfaces;
+
+public interface IHubConnectionFactory
+{
+    IHubConnectionProxy Create(string fullUrl, Func<Task<string?>> accessTokenProvider);
+}
