+++ OpenVPNGateMonitor/Services/DataGateOpenVpnManager/OpenVpnProxy/OpenVpnMicroserviceClient.cs
@@ -4,6 +4,8 @@ using Microsoft.AspNetCore.SignalR.Client;
 using OpenVPNGateMonitor.Hubs;
 using OpenVPNGateMonitor.Models;
 using OpenVPNGateMonitor.Services.Api.Auth.Interfaces;
+using OpenVPNGateMonitor.Services.DataGateOpenVpnManager.OpenVpnProxy.Hubs;
+using OpenVPNGateMonitor.Services.DataGateOpenVpnManager.OpenVpnProxy.Hubs.Interfaces;
 
 namespace OpenVPNGateMonitor.Services.DataGateOpenVpnManager.OpenVpnProxy;
